# mydoc/export/agent-gateway.openapi.yaml
openapi: 3.0.3
info:
  title: VCP Agent Gateway API
  version: 1.0.0
  description: |-
    VCP Agent Gateway 的正式 Native published contract。

    说明：
    - 当前对外协议版本为 `v1`，发布阶段为 `ga`。
    - 该文档覆盖当前正式开放的 `health / capabilities / agents / metrics / memory / context / tools / jobs / events` 资源。
    - 推荐优先使用独立 Gateway 凭证接入；Basic Auth 与 `admin_auth` Cookie 仅作为过渡兼容链路记录。
    - 该契约与 `routes/agentGatewayRoutes.js`、M7-M11 主 specs 以及 contract validation 保持对齐。
  contact:
    name: VCP Team
  x-gateway-version: v1
  x-release-stage: ga
  x-vcp-version: 1.0.0
servers:
  - url: http://localhost:3000
    description: 本地默认端口
  - url: http://localhost:8080
    description: 备用端口
...
security:
  - gatewayKeyHeader: []
  - gatewayBearerAuth: []
  - basicAuth: []
  - adminAuthCookie: []
paths:
...
  /agent_gateway/memory/targets:
    get:
      tags:
        - Memory
      summary: 获取当前 agent 可访问的 memory targets
      operationId: listAgentGatewayMemoryTargets
      parameters:
        - $ref: "#/components/parameters/AgentIdQuery"
        - $ref: "#/components/parameters/RequestIdQuery"
        - $ref: "#/components/parameters/SourceQuery"
        - $ref: "#/components/parameters/RuntimeQuery"
      responses:
        "200":
          $ref: "#/components/responses/MemoryTargetsSuccess"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/memory/search:
    post:
      tags:
        - Memory
      summary: 执行 diary memory 检索
      operationId: searchAgentGatewayMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySearchRequest"
      responses:
        "200":
          $ref: "#/components/responses/MemorySearchSuccess"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/memory/write:
    post:
      tags:
        - Memory
      summary: 写入 durable memory 到指定 diary
      operationId: writeAgentGatewayMemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryWriteRequest"
      responses:
        "200":
          $ref: "#/components/responses/MemoryWriteSuccess"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/context/assemble:
    post:
      tags:
        - Context
      summary: 根据 query 或 recentMessages 组装 recall context
      operationId: assembleAgentGatewayContext
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextAssembleRequest"
      responses:
        "200":
          $ref: "#/components/responses/ContextSuccess"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /agent_gateway/recall/run:
    post:
      tags:
        - Recall
      summary: 执行 gateway recall 并返回完整结果
      operationId: runAgentGatewayRecall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecallRunRequest"
      responses:
        "200":
          $ref: "#/components/responses/RecallRunSuccess"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    gatewayKeyHeader:
      type: apiKey
      in: header
      name: x-agent-gateway-key
      description: 推荐的 Native Gateway 专用凭证。
    gatewayBearerAuth:
      type: http
      scheme: bearer
      description: 与 gateway key 等价的 Bearer 认证输入。
    basicAuth:
      type: http
      scheme: basic
      description: 过渡期兼容的管理后台 Basic Auth。
    adminAuthCookie:
      type: apiKey
      in: cookie
      name: admin_auth
      description: 过渡期兼容的管理态 Cookie。
  headers:
    XRequestId:
      description: 请求 ID
      schema:
        type: string
    XGatewayVersion:
      description: Agent Gateway 运行时版本
      schema:
        type: string
        example: v1
    XTraceId:
      description: Agent Gateway 操作级 trace 标识
      schema:
        type: string
        example: agwop_1234567890abcdef123456
  parameters:
...
    AgentIdQuery:
      name: agentId
      in: query
      required: true
      schema:
        type: string
...
    RequestIdQuery:
      name: requestId
      in: query
      required: false
      schema:
        type: string
...
    SourceQuery:
      name: source
      in: query
      required: false
      schema:
        type: string
    RuntimeQuery:
      name: runtime
      in: query
      required: false
      schema:
        type: string
  responses:
...
    MemoryTargetsSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MemoryTargetsEnvelope"
...
    MemorySearchSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MemorySearchEnvelope"
    MemoryWriteSuccess:
      description: 创建成功或命中幂等去重
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MemoryWriteEnvelope"
    ContextSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ContextEnvelope"
    RecallRunSuccess:
      description: 成功
      headers:
        x-request-id:
          $ref: "#/components/headers/XRequestId"
        x-agent-gateway-version:
          $ref: "#/components/headers/XGatewayVersion"
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RecallRunEnvelope"
...
    InvalidRequest:
      description: 请求参数缺失或格式不正确
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ValidationError:
      description: 请求体校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    TooManyRequests:
      description: 请求超过当前网关运营保护阈值
      headers:
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            rateLimited:
              summary: rate limit rejection
              value:
                success: false
                error: Request rate limit exceeded for this operation
                code: AGW_RATE_LIMITED
                details:
                  operationName: metrics.read
                  traceId: agwop_429example1234567890abcd
                  reason: rate_limited
                  retryAfterMs: 3000
                  limit: 5
                  windowMs: 60000
                meta:
                  requestId: req-metrics-rate-001
                  durationMs: 1
                  gatewayVersion: v1
                  traceId: agwop_429example1234567890abcd
                  operationName: metrics.read
                  retryAfterMs: 3000
    PayloadTooLarge:
      description: 请求体超过当前 operation 的 payload 上限
      headers:
        x-agent-gateway-trace-id:
          $ref: "#/components/headers/XTraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            payloadRejected:
              summary: payload size rejection
              value:
                success: false
                error: Request payload exceeds the configured operation limit
                code: AGW_PAYLOAD_TOO_LARGE
                details:
                  operationName: tool.invoke
                  traceId: agwop_413example1234567890abcd
                  reason: payload_too_large
                  payloadBytes: 16384
                  maxPayloadBytes: 4096
                meta:
                  requestId: req-tool-payload-001
                  durationMs: 0
                  gatewayVersion: v1
                  traceId: agwop_413example1234567890abcd
                  operationName: tool.invoke
                  retryAfterMs: 0
...
    Forbidden:
      description: 被共享 policy 或当前权限限制拒绝
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: 目标 agent、tool、job 或资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    GatewayMeta:
      type: object
      required:
        - requestId
        - durationMs
        - gatewayVersion
      properties:
        requestId:
          type: string
        durationMs:
          type: integer
        gatewayVersion:
          type: string
          example: v1
        authMode:
          type: string
          example: gateway_key
        authSource:
          type: string
          example: x-agent-gateway-key
        gatewayId:
          type: string
          example: vcp-gateway
        traceId:
          type: string
          example: agwop_1234567890abcdef123456
        operationName:
          type: string
          example: tool.invoke
        retryAfterMs:
          type: integer
        toolStatus:
          type: string
          enum:
            - completed
            - accepted
            - waiting_approval
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
        - code
        - meta
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: string
        code:
          type: string
          example: AGW_FORBIDDEN
        details:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    RequestContext:
      type: object
      properties:
        requestId:
          type: string
        sessionId:
          type: string
        agentId:
          type: string
        source:
          type: string
        runtime:
          type: string
          example: native
    AuthContext:
      type: object
      properties:
        requestId:
          type: string
        sessionId:
          type: string
        agentId:
          type: string
        source:
          type: string
        runtime:
          type: string
        adapter:
          type: string
        gatewayId:
          type: string
        authMode:
          type: string
        authSource:
          type: string
        roles:
          type: array
          items:
            type: string
    Message:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
    MemoryTarget:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        type:
          type: string
          example: diary
        allowed:
          type: boolean
...
    MemoryTargetsData:
      type: object
      properties:
        targets:
          type: array
          items:
            $ref: "#/components/schemas/MemoryTarget"
    MemoryTargetsEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/MemoryTargetsData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
...
    MemorySearchRequest:
      type: object
      required:
        - query
        - requestContext
      properties:
        query:
          type: string
        diary:
          type: string
        diaries:
          type: array
          items:
            type: string
        mode:
          type: string
          enum:
            - rag
            - hybrid
            - auto
        k:
          type: integer
        timeAware:
          type: boolean
        groupAware:
          type: boolean
        rerank:
          type: boolean
        tagMemo:
          type: boolean
        profile:
          type: string
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        authContext:
          $ref: "#/components/schemas/AuthContext"
...
    MemorySearchData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MemorySearchItem"
        diagnostics:
          type: object
          additionalProperties: true
    MemorySearchEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/MemorySearchData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    MemoryWriteRequest:
      type: object
      required:
        - target
        - memory
        - requestContext
      properties:
        target:
          type: object
          required:
            - diary
          properties:
            diary:
              type: string
        memory:
          type: object
          required:
            - text
            - tags
          properties:
            text:
              type: string
            tags:
              type: array
              items:
                type: string
            metadata:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
        options:
          type: object
          properties:
            deduplicate:
              type: boolean
              default: true
            idempotencyKey:
              type: string
            bridgeToolName:
              type: string
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        authContext:
          $ref: "#/components/schemas/AuthContext"
    MemoryWriteData:
      type: object
      properties:
        writeStatus:
          type: string
          enum:
            - created
            - skipped_duplicate
        diary:
          type: string
        entryId:
          type: string
        deduplicated:
          type: boolean
        filePath:
          type: string
        idempotentReplay:
          type: boolean
        timestamp:
          type: string
          format: date-time
    MemoryWriteEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/MemoryWriteData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    ContextAssembleRequest:
      type: object
      required:
        - requestContext
      properties:
        query:
          type: string
        recentMessages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        diary:
          type: string
        diaries:
          type: array
          items:
            type: string
        mode:
          type: string
          enum:
            - rag
            - hybrid
            - auto
        maxBlocks:
          type: integer
        tokenBudget:
          type: integer
        maxTokenRatio:
          type: number
          format: float
        minScore:
          type: number
          format: float
        timeAware:
          type: boolean
        groupAware:
          type: boolean
        rerank:
          type: boolean
        tagMemo:
          type: boolean
        profile:
          type: string
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        authContext:
          $ref: "#/components/schemas/AuthContext"
    RecallBlock:
      type: object
      properties:
        text:
          type: string
        metadata:
          type: object
          additionalProperties: true
    ContextData:
      type: object
      properties:
        recallBlocks:
          type: array
          items:
            $ref: "#/components/schemas/RecallBlock"
        estimatedTokens:
          type: integer
        appliedPolicy:
          type: object
          additionalProperties: true
    ContextEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/ContextData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
    RecallRunRequest:
      type: object
      required:
        - agentId
        - query
      properties:
        agentId:
          type: string
        query:
          type: string
        profile:
          type: string
        requestContext:
          $ref: "#/components/schemas/RequestContext"
        authContext:
          $ref: "#/components/schemas/AuthContext"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
...
    RecallRunData:
      type: object
      properties:
        success:
          type: boolean
        agentId:
          type: string
          nullable: true
        profileName:
          type: string
          nullable: true
        requestId:
          type: string
        projectedAt:
          type: integer
        activeProjection:
          type: string
          enum:
            - items
            - recallBlocks
            - fullTextSections
            - attachments
            - hybrid
        items:
          type: array
          items:
            $ref: "#/components/schemas/RecallRunItem"
        recallBlocks:
          type: array
          items:
            $ref: "#/components/schemas/RecallRunBlock"
        fullTextSections:
          type: array
          items:
            $ref: "#/components/schemas/RecallRunFullTextSection"
        attachments:
          type: array
          items:
            type: object
            additionalProperties: true
        diagnostics:
          type: object
          additionalProperties: true
        error:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        status:
          type: integer
    RecallRunEnvelope:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/RecallRunData"
        meta:
          $ref: "#/components/schemas/GatewayMeta"
...